Fix popup page render problem
1) WHY are these changes necessary?
----------------------------------------------------

globals.css is the style base for the entire application. If this file is not introduced, Tailwind CSS class names (e.g. bg-blue-500, text-center) will not take effect, meaning that the application will not display the expected styles and the page may look completely different.

These three line is the key

```
@tailwind base;
@tailwind components;
@tailwind utilities;
```

2) WHAT are these changes?
----------------------------------------------------

Rename the tailwind.css to globals.css for better clarity and import it into popup page

3) WHAT are related problem (if exist)?
----------------------------------------------------

N/A

4) HOW were these changes tested?
----------------------------------------------------

Deployed to dev account